API Reference

Common Tools

BracedExcavation.cauchytoeng — Method

cauchytoeng(strain::Array)

Transform the Cauchy strain tensor to the engineering strain tensor.

Arguments

strain: input strian in the tensor form (3 x 3).

BracedExcavation.engtocauchy — Method

engtocauchy(strain::Array)

Transform the engineering strain tensor to Cauchy strain tensor.

Arguments

strain: input strain in 1D array (1 x 6).

BracedExcavation.interpretElement — Method

interpretElement(elemTags, elemNodeTags)

Reinterpret the Gmsh element data to the Int64 type. Return (a) the total number of elements in one physical group and (b) the nodes array of each element.

Arguments

elemTags: Tags of element in the Gmsh format.

elemNodeTags: Tags of node in the element in the Gmsh format.

BracedExcavation.GP — Type

GP()

Mechanical behaviours at each Gauss point in an element (Global Properties).

Arguments:

E: Young's modulus.

G: Initial shear modulus.

v: Poisson's ratio.

Ks: shear stiffness of interface element.

Kn: normal stiffness of interface element.

H0: deviatoric strain history of previous load step.

H: deviatoric strian history of current load step.

Y: yield flag.

sig: stress of current load step.

sigP: stress of previous load step.

sigG: stress of geostatic state.

eps: elastic strain of current load step.

epsP: elastic strain of previous load step.

BracedExcavation.MESH — Type

MESH()

Geometry properties of the mesh (Global Properties).

Arguments:

lstp: load step.

coord: nodal coordinates.

etpl: element topography.

mat: material names.

ngp: number of Gauss points.

bc: boundary conditions.

krow: global stiffness row indices.

kcol: global stiffness column indices.

kval: global stiffness values.

fint: internal forces.

fb: body forces.

fext: external forces.

react: reaction forces.

uvw: mesh deformation.

BracedExcavation.shapefunc — Method

shapefunc(nD, nen)

Return (a) the total number of Guass points in the elements; (b) the shape function at each Gauss point; (c) the derivation of shape function at each Gauss point; and (d) the weights of Gauss points.

The shape functions of current code contain:

1D element

2-node line element with 1 Gauss point

2D element

4-node rectangle element with 1 Gauss point

6-node triangle element with 3 Gauss points

8-node rectangle element with 4 Gauss points

12-node interface element with 6 Gauss points

3D element

8-node hexahedron element with 8 Gauss points

10-node tetrahedron element with 4 Gauss points

15-node wedge element with 6 Gauss points

Arguments:

nD: dimension of the element

nen: number of nodes in each element

Constitutive models

BracedExcavation.DPconst — Method

DPconst(deps, E, v, frictionangle, cohesion, sigma, strain)

Elasto-perfect plastic Drucker-Prager model.

Arguments:

deps: elastic strain rate, or trial elastic strain rate for the return method.

E: Young's modulus.

v: Poisson's ratio.

frictionangle: friction angle controlling the yield surface.

cohesion: cohession controlling the yield surface.

sigma: stress of previous load step.

strain: strain of previous load step.

BracedExcavation.Hooke3d — Method

Hooke3d(deps, E, v)

Calculate the stress induced by the elastic strain.

Arguments:

deps: elastic strain rate, or trial elastic strain rate for the return method.

E: Young's modulus.

v: Poisson's ratio.

BracedExcavation.elasticD — Method

elasticD(E, v)

Return the elastic stiffness matrix $\mathbf{D}$:

$\mathbf{D} = \frac{E}{(1+\nu)(1-2\nu)} \left[ \begin{array}{cccccc} 1-\nu & \nu & \nu & 0 & 0 & 0 \\ \nu & 1-\nu & \nu & 0 & 0 & 0 \\ \nu & \nu & 1-\nu & 0 & 0 & 0 \\ 0 & 0 & 0 & \frac{1-2\nu}{2} & 0 & 0 \\ 0 & 0 & 0 & 0 & \frac{1-2\nu}{2} & 0 \\ 0 & 0 & 0 & 0 & 0 & \frac{1-2\nu}{2} \end{array} \right].$

Arguments:

E: Young's modulus.

v: Poisson's ratio.

BracedExcavation.smallstrain — Method

smallstrain(deps, H0, G0, v, a, Gam07, direct)

Small strain overlay model (Benz et al. 2009). The shear modulus will be affected by the deviatoric strain history through the Hardin-Drnevich relationships.

Arguments:

deps: elastic strain rate, or trial elastic strain rate for the return method.

H0: deviatoric strain history.

G0: initial shear modulus.

v: Poisson's ratio.

a: parameters for the small-strain overlay model.

Gam07: strain at 70% initial shear stiffness.

direct: direction of the loading: 1 for the monotonic loading, -1 for the unloading/reloading.

BracedExcavation.yieldfunction — Method

yieldfunction(stress, frictionangle, cohesion)

Yield surface for the Drucker-Prage model.

Arguments:

stress: stress at current load step.

frictionangle: friction angle.

cohesion: cohesion.

Mesh generation

BracedExcavation.excavation — Function

excavation(project_name::String, nopopup::Bool)

Generate the unit element mesh by Gmsh with the output: project_name.msh. The size of element is 1 x 1 x 1. The name of the soils is tet_LondonClay, the former means the shape of the element, while the latter means the material property. The name of the diaphragm wall is wed_DiaphragmWall. The props are added during the process of excavation.

Arguments:

project_name: name of the project, corresponding to the name of the mesh.

nopopup: not show the gmsh window to visualise the mesh (default is true).

BracedExcavation.unithex — Function

unithex(project_name::String, nopopup::Bool)

Generate the unit hexahedron element mesh by Gmsh with the output: project_name.msh. The size of element is 1 x 1 x 1. The name of the materials is hex_LondonClay, the former means the shape of the element, while the latter means the material property. Three boundaries are defined as the names of top, sidex, and sidey.

Arguments:

project_name: name of the project, corresponding to the name of the mesh.

nopopup: not show the gmsh window to visualise the mesh (default is true).

BracedExcavation.unittet — Function

unittet(project_name::String, nopopup::Bool)

Generate the unit tetrahedron element mesh by Gmsh with the output: project_name.msh. The size of element is 1 x 1 x 1. The name of the materials is tet_LondonClay, the former means the shape of the element, while the latter means the material property. Three boundaries are defined as the names of top, sidex, and sidey.

Arguments:

project_name: name of the project, corresponding to the name of the mesh.

nopopup: not show the gmsh window to visualise the mesh (default is true).

BracedExcavation.unittetwed — Function

unittetwed(project_name::String, nopopup::Bool)

Generate the unit tetrahedron and wedge element mesh by Gmsh with the output: project_name.msh. The size of element is 1 x 1 x 1. The name of the materials is tet_LondonClay and wed_LondonClay, the former means the shape of the element, while the latter means the material property. Three boundaries are defined as the names of top, sidex, and sidey.

Arguments:

project_name: name of the project, corresponding to the name of the mesh.

nopopup: not show the gmsh window to visualise the mesh (default is true).

BracedExcavation.unitwed — Function

unitwed(project_name::String, nopopup::Bool)

Generate the unit wedge element mesh by Gmsh with the output: project_name.msh. The size of element is 1 x 1 x 1. The name of the materials is wed_LondonClay, the former means the shape of the element, while the latter means the material property. Three boundaries are defined as the names of top, sidex, and sidey.

Arguments:

project_name: name of the project, corresponding to the name of the mesh.

nopopup: not show the gmsh window to visualise the mesh (default is true).

Parameters

BracedExcavation.PARAMSDT — Type

PARAMSDT()

Declare the type of input parameters, and the default values.

Arguments

E0: Young's modulus for the London clay (default: 1.0e6 N/m2).

v0: Poisson's ratio for the London clay (default: 0.2).

E1: Young's modulus for the retaining wall (default: 1.0e9 N/m2).

v1: Poisson's ratio for the retaining wall (default: 0.3).

E2: Young's modulus for the prop (default: 2.0e9 N/m2).

v2: Poisson's ratio for the prop (default: 0.3).

DENS: density (default: 2000.0 kg/m2).

GRAV: gravitational froce (default: -9.81 N/kg).

K0: K0 value (default: 1.0).

W: width of the model (default: 1.0 m).

L: length of the model (default: 1.0 m).

H: height of the model (default: 1.0 m).

P: isotropic consolidation pressure (default: -200e3 N/m2).

U: axial compressure displacement (default: -0.4 m).

LSPTC: load steps for the consolidation (default: 20).

LSTPX: load steps for the compression (default: 20).

CONSTM: constitutive model (default: "Hooked3d").

A: parameters for the small-strain overlay model (default: 0.385).

GAM07: strain at 70% initial shear stiffness (default: 1e-1).

THETA: friction angle for the Drucker-Prager yield surface (default: $\pi/9$).

C: cohesion for the Drucker-Prager yield surface (default: 0.0 N/m2).

KN: normal stiffness for the interface element (default: 1.0e8 N/m2).

KS: shear stiffness for the interface element (default: 1.0e6 N/m2).

THETAI: friction angle for the interface element (default: $\pi/9$).

CI: cohesion for the interface element (default: 0.0 N/m2).

TOL: tolerance for the Newton-Ralphson iteration (default: 1e-9).

Read the gmsh

BracedExcavation.readgmsh_excavation — Method

readgmsh_excavation(argv)

Read the mesh data for the excavation test, need the material physical groups and interface elements.

Arguments:

argv: the input mesh file with suffix, e.g., "BracedExcavation.msh"

BracedExcavation.readgmsh_excavnointer — Method

readgmsh_excavnointer(argv)

Read the mesh data for the excavation test without interface to demonstrate the feasiblity of the code.

Arguments:

argv: the input mesh file with suffix, e.g., "BracedExcavation.msh"

BracedExcavation.readgmsh_triaxial — Method

readgmsh_triaxial(argv)

Read the mesh data for the triaxial test, needs top, sidex, sidey boundaries.

Arguments:

argv: the input mesh file with suffix, e.g., "ElementCode.msh"

Initialise the materials

BracedExcavation.gpmaterial — Method

gpmaterial(mesh, params)

Assign the initial mechanical properties to the Gauss points of elements.

Arguments:

mesh: MESH structure defined in the BracedExcavation.MESH.

params: Input parameters defined by the BracedExcavation.PARAMSDT.

Geostatic state

BracedExcavation.geostatic_excavation — Method

geostatic_excavation(mesh, gps, mesh1, mesh2, dens, grav, K0, H)

Set the geostatic conditions for the braced excavation using the simplified K0 approach. Return the equivalent nodal forces for each stage of excavation.

Arguments:

mesh: MESH structure defined by BracedExcavation.MESH.

gps: GP structure defined by BracedExcavation.GP.

mesh1: mesh information for the first excavation.

mesh2: mesh information for the second excavation.

dens: density of materials.

grav: gravitational forces.

K0: K0 value.

H: height of the model.

BracedExcavation.geostatic_triaxial — Method

geostatic_triaxial(mesh, gps, dens, grav, K0, H)

Set the geostatic conditions for the triaxial tests using the simplified K0 approach.

Arguments:

mesh: MESH structure defined by BracedExcavation.MESH.

gps: GP structure defined by BracedExcavation.GP.

dens: density of materials.

grav: gravitational forces.

K0: K0 value.

H: height of the model.

Boundary conditions

BracedExcavation.boundary_compression — Method

boundary_compression(coord, be2, be3, p, u, lstpx)

Set the boundary conditions for the triaxial tests: the bottom and two sides are roller conditions. The boundaries with names sidex, and sidey are kept constnat with pressure p. During the compression stage, the top boundary is compressed to u within lstpx steps.

Arguments:

coord: coordinates of model nodes.

be2: boundary nodes of sidex.

be3: boundary nodes of sidey.

p: confining pressure p.

u: total displacement of top boundary.

lstpx: load steps for the compression.

BracedExcavation.boundary_consolidation — Method

boundary_consolidation(coord, be1, be2, be3, p)

Set the boundary conditions for the triaxial tests: the bottom and two sides are roller conditions. The boundaries with names top, sidex, and sidey are consolidated to the prescribed pressure p.

Arguments:

coord: coordinates of model nodes.

be1: boundary nodes of top.

be2: boundary nodes of sidex.

be3: boundary nodes of sidey.

p: confining pressure p.

BracedExcavation.boundary_excavation — Method

boundary_excavation(coordinate, W, L)

Set the Dirichlet (displacement) boundary conditions for the excavation model.

Arguments:

coordinate: coordinates of model nodes.

W: width of the model.

L: length of the model.

Loading

BracedExcavation.loading_triaxial — Method

loading_triaxial(mesh, gps, constm, G0, a, Gam07, direct)

Loading step for the FE analysis. All the MESH and GP variables are updated as the global properties.

Arguments:

mesh: MESH struct datatype from BracedExcavation.MESH.

gps: GP struct datatype from BracedExcavation.GP.

constm: constitutive model.

a: parameter for the small strain overlay model.

Gam07: strain at 70% initial shear stiffness.

theta: friction angle for the Drucker-Prager yield surface.

c: cohesion for the Drucker-Prager yield surface.

direct: loading direction, affecting the order the eigen values.

NRtol: tolerance for the Newton_Ralphson iteration.

BracedExcavation.loading_excav — Method

loading_excav(mesh, gps, constm::String, a, Gam07, theta, c, direct, phii, cohi, dens, grav, remesh, NRtol)

Loading steps for the braced excavation. The main difference with the BracedExcavation.loading_triaxial is that the fint, fb, and react induced by the further excavated soils need to be subtracted for the subsequent calculations.

Arguments:

mesh: MESH struct datatype from BracedExcavation.MESH.

gps: GP struct datatype from BracedExcavation.GP.

constm: constitutive model.

a: parameter for the small strain overlay model.

Gam07: strain at 70% initial shear stiffness.

theta: friction angle for the Drucker-Prager yield surface.

c: cohesion for the Drucker-Prager yield surface.

direct: loading direction, affecting the order the eigen values.

phii: friction angle for the interface element.

cohi: cohesion for the interface element.

dens: density of materials.

grav: gravitational forces.

remesh: elements and boundary nodes to be excavated at the next stage.

NRtol: tolerance for the Newton_Ralphson iteration.

BracedExcavation.solve — Method

solve(kval, krow, kcol, f, bc, nd::Int, NRit::Int)

Solve the algebra equations for the global FE analysis.

Arguments:

kval: global stiffness values.

krow: global stiffness row indices.

kcol: global stiffness column indices.

f: right-hand side for the equation $Ku=f$.

bc: boundary conditions.

nd: model dimensions.

NRit: iteration number.

Postprocess

BracedExcavation.etpl_vtk — Method

etpl_vtk(nD, nen)

Transform the nodal numbering in the FE etpl to paraview vtk format.

Arguments:

nD: dimension.

nen: number of nodes in the element.

BracedExcavation.makeVtk — Method

makeVtk(coord, etpl, uvw, meshName)

Generate the mesh data .vtk files.

Arguments:

coord: coordinates of nodes in MESH struct.

etpl: element topography in MESH struct.

uvw: displacement field in the MESH struct.

meshName: output file name.

BracedExcavation.makeVtkGP — Method

makeVtkGP(mpFileName, gpC, gpDat, datNames)

Generate the GP field data in each Gauss point.

Arguments:

mpFileName: output filename.

gpC: total number of Gauss points.

gpDat: stress and strain properties in each Gauss point.

datNames: names of the GP properties.

BracedExcavation.postpro — Method

postpro(uvw, lstp, coord, etpl, gps, outdir)

Generate the mesh vtk files, deformed mesh vtk files, and the properties in each Gauss point.

Arguments:

uvw: displacement field in the MESH struct.

lstp: load step in the MESH struct.

coord: coordinates of nodes in the MESH struct.

etpl: element topography in the MESH struct.

gps: GP properties of elements.

outdir: output directory to save the .vtk files.

BracedExcavation.lineplot_excavation — Method

lineplot_excavation(lstp, uvw, coord, maxnode, outdir)

Plot the wall deflection due to excavation at different excavation stages.

Arguments:

lstp: load step.

uvw: displacement.

coord: coordinates of the model nodes.

maxnode: maximum tag of original nodes.

outdir: directory for saving the figures.

BracedExcavation.lineplot_triaxial — Method

lineplot_triaxial(lstps, constm, results, bc1, lstpc, lstpx, outdir)

Plot the results of triaxial tests: load step and displacement in x, y, z-directions, stress strain relationship, and effective stress paths.

Arguments:

lstps: total load steps.

constm: constitutive models.

results: results data.

bc1: boundary condition for the compression test.

lstpc: load steps for the isotropic consolidation.

lstpx: load steps for the drained compression.

theta: friction angle for the Drucker-Prager yield surface.

c: cohesion for the Drucker-Prager yield surface.

direct: loading direction: 1 for monotonic loading; -1 for unloading/reloading.

outdir: directory for saving the figures.